Reviewed by Oliver Hunt.
        - fix float/int conversions that cause warnings on older versions
          of gcc (like the version currently being used on the PowerPC
          buildbot slave)

        * rendering/RenderPath.cpp:
        (WebCore::RenderPath::paint): Explicitly cast to int.
        (WebCore::RenderPath::addFocusRingRects): Use enclosingIntRect
        to convert a FloatRect to an IntRect.
        * rendering/SVGRootInlineBox.cpp:
        (WebCore::SVGRootInlineBoxPaintWalker::chunkPortionCallback):
        Use 0 for the initial value of an int, rather than 0.0f.
        (WebCore::svgTextStyleForInlineTextBox): Explicitly cast to int.
        (WebCore::applyTextAnchorToTextChunk): Ditto.



Canonical link: https://commits.webkit.org/20542@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@26554 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/rendering/SVGRootInlineBox.cpp
@@ -266,7 +266,7 @@ struct SVGRootInlineBoxPaintWalker {
 
         int textDecorations = styleToUse->textDecorationsInEffect(); 
 
-        int textWidth = 0.0f;
+        int textWidth = 0;
         IntPoint decorationOrigin;
         SVGTextDecorationInfo info;
 
@@ -416,7 +416,7 @@ TextStyle svgTextStyleForInlineTextBox(RenderStyle* style, const InlineTextBox* 
     ASSERT(textBox);
     ASSERT(style);
 
-    return TextStyle(false, xPos, textBox->toAdd(), textBox->m_reversed, textBox->m_dirOverride || style->visuallyOrdered());
+    return TextStyle(false, static_cast<int>(xPos), textBox->toAdd(), textBox->m_reversed, textBox->m_dirOverride || style->visuallyOrdered());
 }
 
 static float cummulatedWidthOrHeightOfTextChunk(SVGTextChunk& chunk, bool calcWidthOnly)
@@ -540,9 +540,9 @@ static void applyTextAnchorToTextChunk(SVGTextChunk& chunk)
 
         // Move target box
         if (chunk.isVerticalText)
-            curBox->setYPos(curBox->yPos() + shift);
+            curBox->setYPos(curBox->yPos() + static_cast<int>(shift));
         else
-            curBox->setXPos(curBox->xPos() + shift);
+            curBox->setXPos(curBox->xPos() + static_cast<int>(shift));
     }
 }
